examples/chat-bot/requirements.txt
olvid-bot>=2.0.0
